refactor: ScoreBoard->ScoreCalculator util 객체 변경
